fix(insights): Fix/refactor click events for bubbles are broken
This fixes (and refactors) all event handlers for release bubbles. Instead of passing them as props to `<BaseChart>`, attach them directly using the echarts intance
