+++ static/app/views/releases/releaseBubbles/useReleaseBubbles.tsx
@@ -1,4 +1,4 @@
-import {type ReactElement, useMemo, useRef} from 'react';
+import {type ReactElement, useCallback, useMemo, useRef} from 'react';
 import {type Theme, useTheme} from '@emotion/react';
 import type {
   CustomSeriesOption,
@@ -6,13 +6,15 @@ import type {
   CustomSeriesRenderItemAPI,
   CustomSeriesRenderItemParams,
   CustomSeriesRenderItemReturn,
+  ElementEvent,
 } from 'echarts';
 import type {EChartsInstance} from 'echarts-for-react';
+import debounce from 'lodash/debounce';
 import moment from 'moment-timezone';
 
 import {closeModal} from 'sentry/actionCreators/modal';
 import {isChartHovered} from 'sentry/components/charts/utils';
-import useDrawer, {type DrawerConfig} from 'sentry/components/globalDrawer';
+import useDrawer from 'sentry/components/globalDrawer';
 import type {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
 import {t, tn} from 'sentry/locale';
 import type {
@@ -20,9 +22,11 @@ import type {
   EChartMouseOutHandler,
   EChartMouseOverHandler,
   ReactEchartsRef,
+  Series,
 } from 'sentry/types/echarts';
 import type {ReleaseMetaBasic} from 'sentry/types/release';
 import {defined} from 'sentry/utils';
+import {trackAnalytics} from 'sentry/utils/analytics';
 import {getFormat} from 'sentry/utils/dates';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
@@ -32,127 +36,26 @@ import {
   BUBBLE_AREA_SERIES_ID,
   BUBBLE_SERIES_ID,
 } from 'sentry/views/releases/releaseBubbles/constants';
-import {createReleaseBubbleHighlighter} from 'sentry/views/releases/releaseBubbles/createReleaseBubbleHighlighter';
 import type {
   Bucket,
   ChartRendererProps,
 } from 'sentry/views/releases/releaseBubbles/types';
 import {createReleaseBuckets} from 'sentry/views/releases/releaseBubbles/utils/createReleaseBuckets';
 
-interface CreateReleaseBubbleMouseListenersParams {
-  alignInMiddle: boolean;
-  buckets: Bucket[];
-  color: string;
-  environments: readonly string[];
-  openDrawer: (
-    renderer: DrawerConfig['renderer'],
-    options: DrawerConfig['options']
-  ) => void;
-  projects: readonly number[];
-  chartRenderer?: (rendererProps: ChartRendererProps) => ReactElement;
+interface LegendSelectChangedParams {
+  name: string;
+  selected: Record<string, boolean>;
 }
 
-/**
- * MouseListeners for echarts. This includes drawing a highlighted area on the
- * main chart when a release bubble is hovered over.
- */
-function createReleaseBubbleMouseListeners({
-  alignInMiddle,
-  chartRenderer,
-  color,
-  openDrawer,
-  buckets,
-  projects,
-  environments,
-}: CreateReleaseBubbleMouseListenersParams) {
-  return {
-    onClick: (params: Parameters<EChartClickHandler>[0]) => {
-      if (params.seriesId !== BUBBLE_SERIES_ID) {
-        return;
-      }
-
-      // `data` is typed as Record<string, any> by ECharts, with no generics
-      // to override
-      const data = params.data as unknown as Bucket;
-
-      // "Full Screen View" for Insights opens in a modal, close before opening
-      // drawer.
-      closeModal();
-
-      openDrawer(
-        () => (
-          <ReleasesDrawer
-            startTs={data.start}
-            endTs={data.final ?? data.end}
-            releases={data.releases}
-            buckets={buckets}
-            projects={projects}
-            environments={environments}
-            chartRenderer={chartRenderer}
-          />
-        ),
-        {
-          shouldCloseOnLocationChange: () => false,
-          ariaLabel: t('Releases drawer'),
-          transitionProps: {stiffness: 1000},
-        }
-      );
-    },
-    onMouseOut: (
-      params: Parameters<EChartMouseOutHandler>[0],
-      instance: EChartsInstance
-    ) => {
-      if (params.seriesId !== BUBBLE_SERIES_ID) {
-        return;
-      }
-
-      // Clear the `markArea` that was drawn during mouse over
-      instance.setOption({
-        series: [{id: BUBBLE_AREA_SERIES_ID, markArea: {data: []}}],
-      });
-    },
-    onMouseOver: (
-      params: Parameters<EChartMouseOverHandler>[0],
-      instance: EChartsInstance
-    ) => {
-      if (params.seriesId !== BUBBLE_SERIES_ID) {
-        return;
-      }
-
-      const data = params.data as unknown as Bucket;
-
-      // Match behavior of ReleaseBubblSeries
-      const xAxisShift = alignInMiddle ? (data.end - data.start) / 2 : 0;
-
-      // Create an empty series that has a `markArea` which is then
-      // rectangular area of the "release bucket" that was hovered over (in
-      // the release bubbles). This is drawn on the main chart so that users
-      // can visualize the time block of the set of relases.
-      instance.setOption({
-        series: [
-          {
-            id: BUBBLE_AREA_SERIES_ID,
-            type: 'custom',
-            renderItem: () => {},
-            markArea: {
-              itemStyle: {color, opacity: 0.1},
-              data: [
-                [
-                  {
-                    xAxis: data.start - xAxisShift,
-                  },
-                  {
-                    xAxis: data.end - xAxisShift,
-                  },
-                ],
-              ],
-            },
-          },
-        ],
-      });
-    },
-  };
-}
+// This needs to be debounced because some charts (e.g. in TimeseriesWidgets)
+// are in a group and share events. Thus on a page with 4 widgets, clicking on
+// a legend item would result in 4 events.
+const trackLegend = debounce((params: LegendSelectChangedParams) => {
+  trackAnalytics('releases.bubbles_legend', {
+    organization: null,
+    selected: Boolean(params.selected.Releases),
+  });
+});
 
 interface ReleaseBubbleSeriesProps {
   alignInMiddle: boolean;
@@ -446,50 +349,256 @@ export function useReleaseBubbles({
     [bubbleSize, totalBubblePaddingY]
   );
 
-  const handleChartRef = (e: ReactEchartsRef | null) => {
-    chartRef.current = e;
+  const buckets = useMemo(
+    () =>
+      (hasReleaseBubbles &&
+        releases?.length &&
+        minTime &&
+        maxTime &&
+        createReleaseBuckets({
+          minTime,
+          maxTime,
+          finalTime: releasesMaxTime,
+          releases,
+          desiredBuckets,
+        })) ||
+      [],
+    [desiredBuckets, hasReleaseBubbles, maxTime, minTime, releases, releasesMaxTime]
+  );
 
-    if (e?.getEchartsInstance) {
-      const echartsInstance = e.getEchartsInstance();
-      createReleaseBubbleHighlighter(echartsInstance, {
-        onLegendChange: (selected: boolean) => {
-          // If `legendSelected` is defined, this hook will assume that the
-          // selected state is "controlled" by the calling component (e.g. it
-          // implements its own event handler and keeps its own legend-selected
-          // state). The hook will return the updated chart options accordingly.
-          if (legendSelected !== undefined) {
-            return;
-          }
-          // Callback for when Releases legend status changes -- we want to
-          // adjust the xAxis/grid accordingly when Releases are visible or
-          // not
-          echartsInstance.setOption({
-            xAxis: selected ? releaseBubbleXAxis : defaultBubbleXAxis,
-            grid: selected ? releaseBubbleGrid : defaultBubbleGrid,
+  const handleChartRef = useCallback(
+    (e: ReactEchartsRef | null) => {
+      chartRef.current = e;
+
+      const echartsInstance: EChartsInstance = e?.getEchartsInstance?.();
+      const highlightedBuckets = new Set();
+
+      const handleMouseMove = (params: ElementEvent) => {
+        // Tracks movement across the chart and highlights the corresponding release bubble
+        const pointInPixel = [params.offsetX, params.offsetY];
+        const pointInGrid = echartsInstance.convertFromPixel('grid', pointInPixel);
+        const series = echartsInstance.getOption().series;
+        const seriesIndex = series.findIndex((s: Series) => s.id === BUBBLE_SERIES_ID);
+
+        // No release bubble series found (shouldn't happen)
+        if (seriesIndex === -1) {
+          return;
+        }
+        const bubbleSeries = series[seriesIndex];
+        const bucketsFromSeries = bubbleSeries?.data;
+
+        if (!bucketsFromSeries) {
+          return;
+        }
+
+        // Try to find the bucket that the mouse is hovered over
+        const bucketIndex = buckets.findIndex(({start, end}: Bucket) => {
+          const ts = pointInGrid[0] ?? -1;
+          return ts >= start && ts < end;
+        });
+
+        // Already highlighted, no need to do anything
+        if (highlightedBuckets.has(bucketIndex)) {
+          return;
+        }
+
+        // If next bucket is not already highlighted, clear all existing
+        // highlights. We also want to clear if bucket was *not* found.
+        if (!highlightedBuckets.has(bucketIndex)) {
+          highlightedBuckets.forEach(dataIndex => {
+            echartsInstance.dispatchAction({
+              type: 'downplay',
+              seriesIndex,
+              dataIndex,
+            });
           });
-        },
-      });
-    }
-  };
+          highlightedBuckets.clear();
+        }
 
-  const buckets =
-    (hasReleaseBubbles &&
-      releases?.length &&
-      minTime &&
-      maxTime &&
-      createReleaseBuckets({
-        minTime,
-        maxTime,
-        finalTime: releasesMaxTime,
-        releases,
-        desiredBuckets,
-      })) ||
-    [];
+        // A bucket was found, dispatch "highlight" action --
+        // this is styled via `renderReleaseBubble` -> `emphasis`
+        if (bucketIndex > -1) {
+          highlightedBuckets.add(bucketIndex);
+          echartsInstance.dispatchAction({
+            type: 'highlight',
+            seriesIndex,
+            dataIndex: bucketIndex,
+          });
+        }
+      };
+      const handleSeriesClick = (params: Parameters<EChartClickHandler>[0]) => {
+        if (params.seriesId !== BUBBLE_SERIES_ID) {
+          return;
+        }
+
+        // `data` is typed as Record<string, any> by ECharts, with no generics
+        // to override
+        const data = params.data as unknown as Bucket;
+
+        // "Full Screen View" for Insights opens in a modal, close before opening
+        // drawer.
+        closeModal();
+
+        openDrawer(
+          () => (
+            <ReleasesDrawer
+              startTs={data.start}
+              endTs={data.final ?? data.end}
+              releases={data.releases}
+              buckets={buckets}
+              projects={projects ?? selection.projects}
+              environments={environments ?? selection.environments}
+              chartRenderer={chartRenderer}
+            />
+          ),
+          {
+            shouldCloseOnLocationChange: () => false,
+            ariaLabel: t('Releases drawer'),
+            transitionProps: {stiffness: 1000},
+          }
+        );
+      };
+
+      const handleMouseOver = (params: Parameters<EChartMouseOverHandler>[0]) => {
+        if (params.seriesId !== BUBBLE_SERIES_ID) {
+          return;
+        }
+
+        const data = params.data as unknown as Bucket;
+
+        // Match behavior of ReleaseBubblSeries
+        const xAxisShift = alignInMiddle ? (data.end - data.start) / 2 : 0;
+
+        // Create an empty series that has a `markArea` which is then
+        // rectangular area of the "release bucket" that was hovered over (in
+        // the release bubbles). This is drawn on the main chart so that users
+        // can visualize the time block of the set of relases.
+        echartsInstance.setOption({
+          series: [
+            {
+              id: BUBBLE_AREA_SERIES_ID,
+              type: 'custom',
+              renderItem: () => {},
+              markArea: {
+                itemStyle: {color: theme.blue400, opacity: 0.1},
+                data: [
+                  [
+                    {
+                      xAxis: data.start - xAxisShift,
+                    },
+                    {
+                      xAxis: data.end - xAxisShift,
+                    },
+                  ],
+                ],
+              },
+            },
+          ],
+        });
+      };
+
+      const handleMouseOut = (params: Parameters<EChartMouseOutHandler>[0]) => {
+        if (params.seriesId !== BUBBLE_SERIES_ID) {
+          return;
+        }
+
+        // Clear the `markArea` that was drawn during mouse over
+        echartsInstance.setOption({
+          series: [{id: BUBBLE_AREA_SERIES_ID, markArea: {data: []}}],
+        });
+      };
+
+      // This fixes a bug where if you hover over a bubble and mouseout via xaxis
+      // (i.e. bottom of chart), the bubble will remain highlighted. This makes it
+      // look buggy and can be misleading especially for bubbles w/ 0 releases.
+      const handleGlobalOut = () => {
+        if (!echartsInstance) {
+          return;
+        }
+
+        const series = echartsInstance.getOption().series;
+        const seriesIndex = series.findIndex((s: Series) => s.id === BUBBLE_SERIES_ID);
+        // We could find and include a `dataIndex` to be specific about which
+        // bubble to "downplay", but I think it's ok to downplay everything
+        echartsInstance.dispatchAction({
+          type: 'downplay',
+          seriesIndex,
+        });
+      };
+
+      const handleLegendSelectChanged = (params: LegendSelectChangedParams) => {
+        if (params.name !== 'Releases' || !('Releases' in params.selected)) {
+          return;
+        }
+        const selected = params.selected.Releases;
+
+        // If `legendSelected` is defined, this hook will assume that the
+        // selected state is "controlled" by the calling component (e.g. it
+        // implements its own event handler and keeps its own legend-selected
+        // state). The hook will return the updated chart options accordingly.
+        if (legendSelected !== undefined) {
+          return;
+        }
+        // Callback for when Releases legend status changes -- we want to
+        // adjust the xAxis/grid accordingly when Releases are visible or
+        // not
+        echartsInstance.setOption({
+          xAxis: selected ? releaseBubbleXAxis : defaultBubbleXAxis,
+          grid: selected ? releaseBubbleGrid : defaultBubbleGrid,
+        });
+
+        trackLegend(params);
+      };
+
+      if (echartsInstance) {
+        /**
+         * MouseListeners for echarts. This includes drawing a highlighted area on the
+         * main chart when a release bubble is hovered over.
+         *
+         * Attach directly to instance to avoid collisions with React props
+         */
+        echartsInstance.on('click', handleSeriesClick);
+        echartsInstance.on('mouseover', handleMouseOver);
+        echartsInstance.on('mouseout', handleMouseOut);
+        echartsInstance.on('globalout', handleGlobalOut);
+        echartsInstance.on('legendselectchanged', handleLegendSelectChanged);
+        echartsInstance.getZr().on('mousemove', handleMouseMove);
+      }
+
+      return () => {
+        if (!echartsInstance) {
+          return;
+        }
+
+        echartsInstance.off('click', handleSeriesClick);
+        echartsInstance.off('mouseover', handleMouseOver);
+        echartsInstance.off('mouseout', handleMouseOut);
+        echartsInstance.off('globalout', handleGlobalOut);
+        echartsInstance.off('legendselectchanged', handleLegendSelectChanged);
+        echartsInstance.getZr().off('mousemove', handleMouseMove);
+      };
+    },
+    [
+      alignInMiddle,
+      buckets,
+      chartRenderer,
+      environments,
+      openDrawer,
+      projects,
+      selection.environments,
+      selection.projects,
+      defaultBubbleGrid,
+      defaultBubbleXAxis,
+      legendSelected,
+      releaseBubbleGrid,
+      releaseBubbleXAxis,
+      theme.blue400,
+    ]
+  );
 
   if (!releases || !buckets.length) {
     return {
       connectReleaseBubbleChartRef: () => {},
-      releaseBubbleEventHandlers: {},
       ReleaseBubbleSeries: null,
       releaseBubbleXAxis: {},
       releaseBubbleGrid: {},
@@ -498,19 +607,6 @@ export function useReleaseBubbles({
 
   return {
     connectReleaseBubbleChartRef: handleChartRef,
-
-    /**
-     * An object map of ECharts event handlers. These should be spread onto a Chart component
-     */
-    releaseBubbleEventHandlers: createReleaseBubbleMouseListeners({
-      alignInMiddle,
-      buckets,
-      chartRenderer,
-      color: theme.blue400,
-      openDrawer,
-      projects: projects ?? selection.projects,
-      environments: environments ?? selection.environments,
-    }),
 
     /**
      * Series to append to a chart's existing `series`
